% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_percentage.R
\name{calculate_percentage}
\alias{calculate_percentage}
\title{Calculate Pellet Reaching Percentages}
\usage{
calculate_percentage(data)
}
\arguments{
\item{data}{A dataframe containing pellet reaching count data.}
}
\value{
A dataframe with percentage columns added.
}
\description{
This function calculates percentage-based performance metrics.
It assumes input data contains raw count columns.
}
\examples{
# example code
df <- data.frame(
   "First trial success" = c(10, 12),
  "Total grasps" = c(20, 20),
  "Total success" = c(15, 18),
  "Total fail" = c(5, 2),
  "Total pellet to box" = c(14, 16),
  check.names = FALSE
)

calculate_percentage(df)
}
