% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_check_required_columns.R
\name{helper_check_required_columns}
\alias{helper_check_required_columns}
\title{Check Required Columns in a Dataframe}
\usage{
helper_check_required_columns(df, required_cols)
}
\arguments{
\item{df}{A dataframe to check.}

\item{required_cols}{A character vector of column names that must be present in df.}
}
\value{
Returns TRUE if all required columns are present. Stops with an error otherwise, and gives the missing columns.
}
\description{
This helper function checks whether a dataframe contains all required columns.
If any columns are missing, the function stops and returns an error message.
}
\examples{
 df <- data.frame(a = 1, b = 2)
helper_check_required_columns(df, c("a", "b")) # returns TRUE
\dontrun{
helper_check_required_columns(df, c("a", "c")) # would throw an error
}
}
