% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_long_format.R
\name{make_long_format}
\alias{make_long_format}
\title{Convert Pellet Reaching Data to Long Format}
\usage{
make_long_format(data)
}
\arguments{
\item{data}{A dataframe with percentage columns.}
}
\value{
A long-format dataframe for plotting
}
\description{
Convert Pellet Reaching Data to Long Format
}
\examples{
\dontrun{
df <- data.frame(
  Rat_number = c(1,2),
  Blinding_group = c("A","B"),
  First_trial_success_pct_Baseline = c(50, 60),
  First_trial_success_pct_Post = c(70, 80),
  Total_success_pct_Baseline = c(60, 65),
  Total_success_pct_Post = c(75, 85),
  Total_fail_pct_Baseline = c(10, 5),
  Total_fail_pct_Post = c(5, 2),
  Total_pellet_to_box_pct_Baseline = c(55, 60),
  Total_pellet_to_box_pct_Post = c(80, 90)
)

df_long <- make_long_format(df)
head(df_long)
}
}
