% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot Pellet Reaching Results}
\usage{
plot_results(data_long, sig_labels)
}
\arguments{
\item{data_long}{Long-format dataframe from \code{make_long_format()}.}

\item{sig_labels}{A dataframe with p-values and significance labels.}
}
\value{
A combined ggplot object.
}
\description{
Plot Pellet Reaching Results
}
