% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_pdf.R
\name{plot_to_pdf}
\alias{plot_to_pdf}
\title{Save a ggplot object to PDF}
\usage{
plot_to_pdf(plot_object, filename, width = 14, height = 12, dpi = 300)
}
\arguments{
\item{plot_object}{A ggplot object}

\item{filename}{Name of the output file}

\item{width}{Width of the PDF in inches (default 14)}

\item{height}{Height of the PDF in inches (default 12)}

\item{dpi}{Resolution of the PDF in dots per inch (default 300)}
}
\value{
Invisubly returns th filename of the saved PDF
}
\description{
This function saves a ggplot object to a PDF file with customizable dimensions and resolution.
}
\examples{
#' \dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
plot_to_pdf(p, "mtcars_plot.pdf")
}

}
